+++ apps/shared/avifjpeg.h
@@ -10,6 +10,11 @@
 extern "C" {
 #endif
 
+// Decodes the jpeg file at path 'inputFilename' into 'avif'.
+// 'ignoreGainMap' is only relevant for jpeg files that have a gain map
+// and only if AVIF_ENABLE_EXPERIMENTAL_JPEG_GAIN_MAP_CONVERSION is ON
+// (requires AVIF_ENABLE_EXPERIMENTAL_GAIN_MAP and libxml2). Otherwise
+// it has no effect.
 avifBool avifJPEGRead(const char * inputFilename,
                       avifImage * avif,
                       avifPixelFormat requestedFormat,
@@ -17,8 +22,14 @@ avifBool avifJPEGRead(const char * inputFilename,
                       avifChromaDownsampling chromaDownsampling,
                       avifBool ignoreColorProfile,
                       avifBool ignoreExif,
-                      avifBool ignoreXMP);
+                      avifBool ignoreXMP,
+                      avifBool ignoreGainMap);
 avifBool avifJPEGWrite(const char * outputFilename, const avifImage * avif, int jpegQuality, avifChromaUpsampling chromaUpsampling);
+
+#if defined(AVIF_ENABLE_EXPERIMENTAL_JPEG_GAIN_MAP_CONVERSION)
+// Parses XMP gain map metadata. Visible for testing.
+avifBool avifJPEGParseGainMapXMP(const uint8_t * xmpData, size_t xmpSize, avifGainMapMetadata * metadata);
+#endif
 
 #ifdef __cplusplus
 } // extern "C"